sdk: Add built-in support to some chains
